types: Update Len of Array and Vector type from int to int64.

+++ ir/types/derived.go
@@ -173,12 +173,12 @@ type VectorType struct {
 	// Element type.
 	elem Type
 	// Vector length.
-	len int
+	len int64
 }
 
 // NewVector returns a new vector type based on the given element type and
 // vector length.
-func NewVector(elem Type, len int) *VectorType {
+func NewVector(elem Type, len int64) *VectorType {
 	return &VectorType{elem: elem, len: len}
 }
 
@@ -206,7 +206,7 @@ func (t *VectorType) Elem() Type {
 }
 
 // Len returns the length of the vector type.
-func (t *VectorType) Len() int {
+func (t *VectorType) Len() int64 {
 	return t.len
 }
 
@@ -220,12 +220,12 @@ type ArrayType struct {
 	// Element type.
 	elem Type
 	// Array length.
-	len int
+	len int64
 }
 
 // NewArray returns a new array type based on the given element type and array
 // length.
-func NewArray(elem Type, len int) *ArrayType {
+func NewArray(elem Type, len int64) *ArrayType {
 	return &ArrayType{elem: elem, len: len}
 }
 
@@ -253,7 +253,7 @@ func (t *ArrayType) Elem() Type {
 }
 
 // Len returns the length of the array type.
-func (t *ArrayType) Len() int {
+func (t *ArrayType) Len() int64 {
 	return t.len
 }
